+++ Sources/Module/Channel/NotificationChannel/Feed/SBUFeedNotificationChannelModule.List.swift
@@ -507,7 +507,6 @@ extension SBUFeedNotificationChannelModule {
                     )
                 }
             }
-
             
         }